+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/Image.tsx
