+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/List.tsx
