+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Blocks/tests/Quote.test.tsx
