+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/BlocksEditor.tsx
@@ -9,13 +9,13 @@ import { withHistory } from 'slate-history';
 import { type RenderElementProps, Slate, withReact, ReactEditor, useSlate } from 'slate-react';
 import styled from 'styled-components';
 
-import { codeBlocks } from './blocks/Code';
-import { headingBlocks } from './blocks/Heading';
-import { imageBlocks } from './blocks/Image';
-import { linkBlocks } from './blocks/Link';
-import { listBlocks } from './blocks/List';
-import { paragraphBlocks } from './blocks/Paragraph';
-import { quoteBlocks } from './blocks/Quote';
+import { codeBlocks } from './Blocks/Code';
+import { headingBlocks } from './Blocks/Heading';
+import { imageBlocks } from './Blocks/Image';
+import { linkBlocks } from './Blocks/Link';
+import { listBlocks } from './Blocks/List';
+import { paragraphBlocks } from './Blocks/Paragraph';
+import { quoteBlocks } from './Blocks/Quote';
 import { BlocksContent } from './BlocksContent';
 import { BlocksToolbar } from './BlocksToolbar';
 import { withLinks } from './plugins/withLinks';
